Conflicts:
	src/main/resources/org/ecocean/package.jdo
	src/main/webapp/WEB-INF/web.xml
	src/main/webapp/encounters/encounter.jsp
	src/main/webapp/images/masthead.jpg

+++ src/main/webapp/appadmin/admin.jsp
@@ -207,9 +207,24 @@
             Shepherd Project instances that have transient data (i.e., the Shepherd Project is not the primary data store).</em>
             </p>
 
-            <form name="deleteAll" method="post" action="../DeleteAllDataPermanently">
+            <form onsubmit="return confirm('Are you sure you want to delete all encounters? WARNING! This will cause complete data loss!');" name="deleteAll" method="post" action="../DeleteAllDataPermanently">
 
               <input name="deleteAllData" type="submit" id="deleteAllData" value="Delete All Data PERMANENTLY">
+              </p></form>
+          </td>
+        </tr>
+      </table>
+      
+        <p>&nbsp;</p>
+      <table width="600" border="1">
+        <tr>
+          <td>
+            <p><font size="+1">Reset Common Configuration Properties</font></p>
+            <p></p>Reset the values in the commonConfiguration.properties file.</p>
+
+            <form name="ResetCommonConfigurationForm" method="post" action="../ResetCommonConfiguration">
+
+              <input name="ResetCommonConfiguration" type="submit" id="ResetCommonConfiguration" value="Reset">
               </p></form>
           </td>
         </tr>
